LayoutTests/platform/glib/fast/scrolling/scroll-to-anchor-zoomed-header-expected.txt

Tests scrolling to an anchor in a zoomed page with a header scrolls to the correct location.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL document.scrollingElement.scrollTop should be 1019. Was 1020.
PASS document.scrollingElement.scrollLeft is 0
PASS successfullyParsed is true

TEST COMPLETE
Anchor is here
